BlogCreation-AIAgent/Scripts/requirements.txt
crewai
crewai_tools
load_dotenv
langchain_google_genai
youtube-transcript-api
yt_dlp==2023.11.14 
youtube-transcript-api==0.6.1